+++ doc/RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 ## Version 0.8.3 ##
 
 * Issue #7: duplicate ODN detector can be fooled by inconsistent case
+* Issue #9: allow blank lines in blacklist files
 * Issue #10: shorten class names on X.509 validators
 * Improved error status messages from `X509RSAOpenSSLBlacklistValidator`.